[FIX] hit판정을 ontriger에서 적용시키는 것으로 변경, 명중 회피와 공격 방어도 적용

+++ Assets/Scripts/Interface/IDamageable.cs
@@ -8,7 +8,7 @@ namespace Interface
     [Serializable]
     public struct DamageData
     {
-        public int Damage;
+        public float Damage;
         public Vector3 KbForce;
         public ElementalType ElementalType;
